security: implement strict tutor-child authorization checks in HTTP controllers and WebSocket gateway

+++ safesteps-backend-master/src/usuarios/tutor.controller.ts
@@ -7,6 +7,7 @@ import {
   Param,
   Delete,
   UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 import { TutorService } from './tutor.service';
 import { CreateTutorDto } from './dto/create-tutor.dto';
@@ -66,32 +67,50 @@ export class TutorController {
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  findOne(@Param('id') id: string, @GetUser() user: User) {
+    if (user.id !== +id) {
+      throw new ForbiddenException('No tienes permiso para ver los datos de este tutor');
+    }
     return this.tutorService.findOne(+id);
   }
 
   @Get(':id/hijos')
-  getHijos(@Param('id') id: string) {
+  getHijos(@Param('id') id: string, @GetUser() user: User) {
+    if (user.id !== +id) {
+      throw new ForbiddenException('No tienes permiso para ver los hijos de este tutor');
+    }
     return this.tutorService.getHijos(+id);
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateTutorDto: UpdateTutorDto) {
+  update(@Param('id') id: string, @Body() updateTutorDto: UpdateTutorDto, @GetUser() user: User) {
+    if (user.id !== +id) {
+      throw new ForbiddenException('No tienes permiso para actualizar este tutor');
+    }
     return this.tutorService.update(+id, updateTutorDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Param('id') id: string, @GetUser() user: User) {
+    if (user.id !== +id) {
+      throw new ForbiddenException('No tienes permiso para eliminar este tutor');
+    }
     return this.tutorService.remove(+id);
   }
 
   @Post(':tutorId/hijos/:hijoId')
-  addHijo(@Param('tutorId') tutorId: string, @Param('hijoId') hijoId: string) {
+  addHijo(@Param('tutorId') tutorId: string, @Param('hijoId') hijoId: string, @GetUser() user: User) {
+    if (user.id !== +tutorId) {
+      throw new ForbiddenException('No tienes permiso para vincular hijos a este tutor');
+    }
     return this.tutorService.addHijo(+tutorId, +hijoId);
   }
 
   @Delete(':tutorId/hijos/:hijoId')
-  removeHijo(@Param('tutorId') tutorId: string, @Param('hijoId') hijoId: string) {
+  removeHijo(@Param('tutorId') tutorId: string, @Param('hijoId') hijoId: string, @GetUser() user: User) {
+    if (user.id !== +tutorId) {
+      throw new ForbiddenException('No tienes permiso para desvincular hijos de este tutor');
+    }
     return this.tutorService.removeHijo(+tutorId, +hijoId);
   }
 }
